+++ lamindb/_feature_set.py
@@ -79,11 +79,10 @@ def __init__(self, *args, **kwargs):
 @doc_args(FeatureSet.save.__doc__)
 def save(self, *args, **kwargs) -> None:
     """{}"""
-    logger.info("Saving FeatureSet...")
     super(FeatureSet, self).save(*args, **kwargs)
     if hasattr(self, "_features"):
         related_name, records = self._features
-        bulk_create(records, logging=False)
+        bulk_create(records)
         getattr(self, related_name).set(records)